+++ flowapp/models/user.py
@@ -1,4 +1,5 @@
-from sqlalchemy import event
+from typing import List, Optional
+from sqlalchemy import event, select
 from .base import db, user_role, user_organization
 from .organization import Organization
 
@@ -46,16 +47,28 @@ class User(db.Model):
             self.organization.remove(org)
 
         for role_id in form.role_ids.data:
-            my_role = db.session.query(Role).filter_by(id=role_id).first()
+            my_role = db.session.execute(select(Role).filter_by(id=role_id)).scalar_one()
             if my_role not in self.role:
                 self.role.append(my_role)
 
         for org_id in form.org_ids.data:
-            my_org = db.session.query(Organization).filter_by(id=org_id).first()
+            my_org = db.session.execute(select(Organization).filter_by(id=org_id)).scalar_one()
             if my_org not in self.organization:
                 self.organization.append(my_org)
 
         db.session.commit()
+
+    @classmethod
+    def get_all(cls) -> List["User"]:
+        return db.session.scalars(select(cls)).all()
+
+    @classmethod
+    def get_all_ordered(cls) -> List["User"]:
+        return db.session.scalars(select(cls).order_by(cls.name)).all()
+
+    @classmethod
+    def get_by_uuid(cls, uuid: str) -> Optional["User"]:
+        return db.session.scalars(select(cls).filter_by(uuid=uuid)).first()
 
 
 class Role(db.Model):
@@ -69,6 +82,10 @@ class Role(db.Model):
 
     def __repr__(self):
         return self.name
+
+    @classmethod
+    def get_all_ordered(cls) -> List["Role"]:
+        return db.session.scalars(select(cls).order_by(cls.name)).all()
 
 
 # Event listeners for Role